update tracing crates
This branch updates the proxy to depend on 0.1 versions of `tracing-log`
and `tracing-subscriber`, rather than alpha versions. The dependency on
`tracing-fmt` was removed, as that code was moved into
`tracing-subscriber` in tokio-rs/tracing#311.

Additionally, this should now allow us to use filter configuration that
involve dynamic filters on span fields, as well as on static targets.
Note that this may be of limited value *currently*, since the proxy
still uses our sui generis logging contexts rather than `tracing` spans,
but could be useful in the future.

Note that this is a duplicate of #342. I've submitted a new PR since CI
seems to be stuck refusing to build that branch. Therefore, this
closes #342.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>