ci: Use a readymade release action
The first pass at releases used a container I found on the internet
called `github/ghx`, which provides a CLI interface to github that was
usable within actions. I've contacted the author, who says it is
unmaintained (and the sources for it are not public); so I've elected to
use softprops' release action.

In order to do this, I added a local action that handles the parsing of
the git `refs/*` syntax into the proper release metadata.