internal: Remove the SetKeepalive trait
In preparation of decoupling the listener logic from TLS discovery, I'm
trying to simplify the transport module.

This change is intended to make the `listen` and `connect` modules
primarily responsible for preparing a `TCPStream` before any other
transformations are applied to the IO stream. The `internal::Io` trait
can be narrowed and the various `Io` impls no longer need to proxy this
API.

In follow-up changes, the `listen` module will be decoupled from
identity/TLS; and some of the `transport` core types will be split out
into subcrates.