internal: Split transport utilities into a subcrate
The `transport` module contains utilities for working with transports
that are not otherwise coupled to the main application code.

This change replaces the `WrapServerTransport` trait with a
`TransportLabels` trait that is only responsible for producing metrics
labels. The transport module is otherwise (mostly) ignorant of transport
labels. (It is responsible for setting an Errno label).

`proxy::Server` has (temporarily) been tasked with instrumenting the
connection with telemetry, though this will likely be further decoupled
in follow-up changes.