Remove the LINKERD2_PROXY_INBOUND_FORWARD config
The `LINKERD2_PROXY_INBOUND_FORWARD` configuration was implemented to
support local testing; but the proxy has evolved to the point where it
needs a proper runtime harness (as in docker or the integration test
infrastructure) to run.

This option is just a bit of debt that should be removed to simplify
code/diagnostics.

Closes linkerd/linkerd2#3609