Modularize proxy initialization (#382)
The proxy has classically had one very large function responsible for
wiring together the entire proxy. Recently, we've split some of the
logic into inbound and outbound crates, which helps compile times and
modularity.

This change continues this work, moving all of the "front-end"
configuration logic into the `linkerd-app` crate. Configuration is
structured in terms of a struct for each module; and a single `env
module is responsible for building an application. This opens the door
for additional configuration frontends, as long as they are able to
build an `app::Config`.