Update tower-balance
tower-balance has been refactored and updated to use a new version of
the `rand` crate.

This change udpates the project to use the latest tower-balance, rand,
and quickcheck dependencies

I've soak-tested this over the weekend and the new balancer
implementation is at least as good as the prior version in terms of
latency and CPU usage.