profiles: Eagerly read profiles off the wire
We believe that the destination's connection can become stalled because
the dst client's connection window is exhausted. One potential cause
for such a stall is the manner in which the destination client buffers
profile updates such that profile updates may not be read for idle
routes.

This change modifies the profile client to eagerly read updates from the
wire, publishing the newest profile to a `watch`.