router: Ensure that the purge task completes (#396)
The router's cache purging tasks never complete. This means that
when a router is dropped, it's purge task is leaked into the executor,
possibly holding services.

This change uses an mpsc to ensure that the purge task is notified when
all of the routers using this cache have been dropped.