Configure the HTTP/2 connection window to 1MB
Currently, the proxy uses the protocol defaults for HTTP/2 connection
window size (64K). Because this is the same as the stream window size, a
single stream can consume the entire connection window, stalling other
streams.

This change alters the default connection window to be 1MB, supporting
up to 16 stalled streams.

While we primarily care about this setting for client (response)
receiver windows, it seems like a good thing to open the server
(request) receiver window as well (and the configuration is not
currently distinguishable).