Remove clone in opencensus exporter to ensure task is notified (#405)
In the opencensus span exporter, we call `poll_ready` on the span send channel sender to determine if there is capacity in this channel to send spans.  However, the channel sender that we `poll_ready` on is a clone of the original and the clone is dropped shortly after the call to `poll_ready`.  This can cause our interest registration to be dropped and the task may not be notified when capacity becomes available.

The clone is unnecessarily and we can remove it.

Signed-off-by: Alex Leong <alex@buoyant.io>

Signed-off-by: Alex Leong <alex@buoyant.io>