internal: Split retry, http-classify, and http-metrics
The retry middleware is coupled to both the HTTP metrics infrastructure
as well as the various request extensions that are required to route a
request (i.e. via `TryClone`). Because of this, retry metrics are
present for all HTTP metric scopes, even where retries can't apply.

This change extracts several components from the `proxy-http`
conglomerate subcrate:

* `http-classify` includes classification traits and a middleware layer;
* `http-metrics` wraps `tower::retry` with a `MakeService` configuration layer;
* `retry` wraps `tower::retry` with a `MakeService` configuration layer;
* `app-core` now handles retry-configuration

Additionally, the `metrics::Metric` type is now generic over the metric
name type so that prefixed names do not need to be formatted eagerly.

The `route_actual_retry_skipped_total` counter has been replaced with a
`route_retryable_total` counter. This new value is only present when
retries are enabled on a route; and a `skipped` labels is added when
retryability is ignored (i.e. due to budget being exhausted). This
metric is not consumed by the control plane, so it should be safe to
change.