trace: update `tracing-subscriber` to 0.2.0-alpha.4
This commit updates the proxy's `tracing-subscriber` dependency from
0.1.4 to 0.2.0-alpha.4. This alpha includes tokio-rs/tracing#514, which
fixes a pair of issues which prevent spans from being considered closed.
One of these issues exists in 0.1.x as well.

When spans are not considered closed, storage allocated for those spans'
per-span data is not reused. Therefore, this issue results in a memory
leak.

Note that some span storage will still not appear to be
_deallocated_ after this change, but that storage will be reused when
spans are allowed to close correctly.

In addition, I've updated the proxy's `trace` module to remove the
custom formatter. This was used only to add Linkerd's custom logging
contexts to `tracing-subscriber` logs, and as we no longer use the
logging contexts, we can use `tracing-subscriber`'s default formatter.
Now, we only override the timestamp format.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>