router: Avoid large up-front allocations
We initially opted to pre-allocate the delay queue to avoid allocation
at runtime. However, practically, router capacities are _much_ higher
than their typical usage; so we incur an extra ~1MB+ of allocation that
will never really be used (for outbound routers).

This change initializes the router's DelayQueue with only a single slot
by default. The capacity will be increased at runtime as needed.