discover: Warn on discovery error
Users have reported encountering the "discovery task failed" error
message. Unfortunately, the underlying error is only logged at the debug
level, so we can't easily diagnose the underlying issue.

This change raises the logging level of these errors to be WARN so that
they are visible at the default logging levels.