fallback: Operate on Services instead of Layers
The fallback middleware currently operates by applying a primary layer
to an underlying service and, if that primary service does not produce a
result, the fallback layer is applied to the underlying service to
produce a result.

This change modifies the middleware to no longer require that the
underlying service be shared by the primary and fallback layers.
Instead, the middleware attempts to produce a result from the underlying
service and, if that fails, the fallback service is used.

This will allow the fallback middleware to be used more flexibly in
upcoming changes.