timeout: Do not synthesize HTTP response
The `proxy::http::timeout` middleware currently creates HTTP responses
when a timeout is encountered. Instead, we should allow the error to
propagate through the stack as an error so that all synthesized
responses happen in one place.

This change means that we no longer annotate a `status_code` label in
the `route_response_total` metric, though the `classification` and
`error` labels remain unchanged.

We can also avoid using a `tower::util::Either` by encoding the optional
behavior within the implementation (cutting down on compile-time type
bloat).