inbound: Box client transports to ensure shutdown (#445)
The inbound client may never initiate a TLS connection, but we use a TLS
middleware to ensure the underlying transport is boxed. The BoxedIo type
enforces shutdown semantics differently than the raw TcpStream, and we
rely on this behavior in TCP forwarding.

This change introduces a stack::map_response middleware; and replaces
use of the inbound TLS client with a layer that boxes transports.