server: Simplify HTTP server type constraints
`proxy::server` accepts a `MakeService` to build an HTTP service for
each connection; but it doesn't properly drive the service before
invoking it (which is a violation of the Service contract, though a
benign one in this instance).

This change uses `stack::NewService` to avoid these readiness concerns.
We also relax unnecessary type constraints.