orig-proto: Be more flexible to stack placement (#444)
The orig-proto Upgrade middleware unnecessarily captures a PhantomData;
and does URI normalization even though we have another dedicated
middleware for that.

This change modifies orig-proto::Upgrade to use the stack configuration
to determine whether the request's URI was absolute (these requests may
not share a Hyper client, so they need to be on different stacks). This
module no longer normalizes URIs.

The inbound::orig_proto_downgrade module is removed, as it mostly
boilerplate.