timeout: Introduce FailFast, Idle, and Probe middlewares (#452)
This change introduces two new timeout middlewares: `idle` and
`failfast`.

`idle` causes the service to start failing if polled after being ready
and unused for some timeout. This is intended to be driven by
`probe-buffer`.

`failfast` causes the service to become ready, failing requests, if the
inner service does not become ready within a timeout.