cache: Let services self-evict
This change modifies the `cache` crate to remove its background task
which is responsible for tracking service use and evicting idle
services.

Now, the cache (1) has no capacity limit and (2) evicts services from
the cache when they drop their underlying service.

This is accomplished by modifying the `CacheLayer` to accept a layer
that is responsible for making the inner service shareable and for
dropping its inner service when it should be considered defunct. This is
accomplished by moving an `Arc<()>` into the inner stack and holding a
`Weak<()>` reference to it in the cache module. Then, before a request
is processed by the cache, it attempts to upgrade each cache entry and
drops services that have dropped the inner service (holding the strong
reference).