Replace ExtractAuthority with CanOverrideAuthority
The extractor pattern is useful when we need to use additional
configuration or when we don't control the type over which we're
extracting a value (like an http::Request). But in the case of authority
overrides, we have all of the information we need on the target type.

This change replaces the ExtractAuthority strategy with a
CanOverrideAuthority trait that must be implemented by targets.

I've also modified logging to be at the debug level (request
modification seems useful to observe when debugging), and I've renamed
some variables to match the tracing output.

I think, as a follow up, I'd prefer that we call this module
`override_authority` (rather than "overwrite"), as this matches the
naming in the controller/API.

I've also moved `layer => Layer::new`, which is the newer idiom we're
targetting.