Handle GRPC body errors (#493)
When gRPC implementations encounter an underlying transport error, they
surface the error via an `UNAVAILABLE` gRPC status code. Linkerd breaks this
behavior by converting these errors to stream resets, which instead appear as
an `UNKNOWN` error to applications.

This change modifies the proxy's error handling logic to set the the correct gRPC
status code when a stream fails with a hyper error, which should work more
seamlessly with gRPC error handling. This status is only set when the request's
content-type indicates the stream is sending gRPC messages.