update linkerd2-cache and linkerd2-lock to std::future (#490)
This branch updates the `linkerd2-cache` crate (and `linkerd2-lock`,
which it relies on), to use `std::future`.

Unlike previous PRs, the `linkerd2-lock` update is a more substantial
rewrite, because upstream API changes made the previous implementation
no longer possible. In particular, `tokio::sync::Mutex` does not provide
a `poll_acquire` method the way that `tokio_sync::Lock` did in Tokio
0.1. The removal of `poll_acquire` is largely due to the rewrite of
Tokio's synchronization primitives to use an intrusive linked list-based
semaphore (tokio-rs/tokio#2325), which requires waiters to be pinned to
ensure safety of the intrusive list. Allowing permits to be acquired
only from a future ensures correct pinning of waiters.

Therefore, I've implemented a new `Lock` for Linkerd which uses a Tokio
`Mutex` internally. The `Lock` owns a boxed instance of the future returned
by `Mutex::lock_owned`, and drives it in `poll_acquire`. This allows adapting
the poll-based Tower interface with the futures-only, pinned interface.