buffer: Move idle timeouts into the buffer (#502)
The layered approach to idle timeouts does not behave well with the
buffer. When the buffer's dispatch receives a request exactly as the
timeout expires, it polls the inner service (and therefore the idle
timeout) before polling for new requests, causing the requests to be
failed with an idle error.

This change removes the `idle` timeout layer, instead opting to do
idle-based timeouts in the buffer directly. This allows us to poll for
new requests before checking idleness, eliminating the possibility of
this kind of data-race.

Addresses linkerd/linkerd2#4274