Buffer requests while the service is pending
The buffer uses a watch to advertise readiness to consumers. This is
intended to:

1. Prevent the out service from becoming ready before the inner service
   has initialized.
2. Surface errors to outer services' `poll_ready`.

It, however, currently advertises when the inner service moves into
pending, preventing requests from being admitted to the buffer, and then
basically making consumers race to enter the buffer.

This change also relaxes its use of the ready watch channel, especially
to signal that the task should complete. Instead, the dispatch task
completes as soon as all requests have been processed.