port stack-metrics to std::future (#513)
I had forgotten that `linkerd2-stack-metrics` is actually kind of load
bearing: the `TrackServiceLayer` is necessary for powering cache
eviction as well as being used for debugging. It turns out that we need
this to create stacks that work.

This branch ports `linkerd2-stack-metrics` to std::future. This change
was pretty trivial; I just updated to use the new `tower::Service`
trait. I also pruned a bunch of dependencies that this crate wasn't
actually using from the Cargo.toml.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>