turn inbound proxy back into a proxy (#516)
Now that the HTTP connect- and accept-side machinery has all been
updated to use `std::future`, we can start turning the proxy back
into...a proxy. This branch does so on the inbound side. Much of the
actual functionality (approximately everything not directly required to
actually proxy requests) is still disabled, as a lot of the code still
needs to be updated.

I had to make a few additional changes in order to get the proxy to
compile, primarily to deal with trait bounds on `Service` impls that
were no longer satisfied as a result of changes made to update other
parts of the codebase. I also disabled two more of the  `transparency`
integration tests, both of which tested whether the proxy stripped
connection-level headers from proxied requests. Previously, these were
passing "accidentally" — because the proxy was replaced with a HTTP
server that just responded to every request with an empty 200 OK
response, the headers were, of course, not present. Now that we're
actually proxying requests, these tests fail, because the code for
stripping headers has yet to be updated.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>