update dns to std::future (#520)
This branch updates the `linkerd2-proxy-dns` crate to `std::future` and
the latest release of `trust-dns-resolver`, and updates the uses of DNS
in `linkerd2-proxy-app` to track the changes. 

For the most part, this change is pretty straightforward. However, the
`std::future` version of `trust-dns-resolver` changed how the background
task driving DNS resolution is executed. Previously, it was returned by
the `AsyncResolver` constructor, but now, the constructors are `async
fn`s that implicitly spawn the background task on the runtime that
executes them.

Therefore, I had to change how the DNS resolver is constructed. I opted
to build the admin runtime early and have it block on the DNS resolver
construction, and then pass that runtime to the admin thread when it is
spawned. This seemed better than making the DNS resolver lazy by making
the underlying `trust-dns` async resolver an `Option` or something, and
having it panic when the resolver hasn't been constructed yet. Doing
that safely would have required adding a `RwLock` or something, which
didn't seem ideal.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>