fix subtle change in test client behavior
This is a fun one: the test `retry_reconnect_errors` relied on
`Client::request_async` *eagerly* causing the test client to connect to
the proxy when called, rather than when the returned future is polled.
Switching `request_async` to an `async fn` broke this property — the
*entire* body of the function, even the synchronous parts, was run only
when the `async fn` future is polled. This means that the connection was
not opened when the test expected it to be, and the subsequent assertion
that it was open always failed.

This commit changes this back into a synchronous function returning
`impl Future`. This is identical at the callsite (it still returns
something that you can `.await`), but calling the function now once
again eagerly tells the test client background thread to open a
connection. This fixes the test.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>