update reconnect to std::future (#531)
This branch updates the `reconnect` middleware to `std::future`.
Updating reconnect was pretty straightforward and mechanical, so it
should be easy to review. 

Most of the complexity was actually in getting the test
`retry_reconnect_errors` which exercises this functionality to pass: the
test uses transport metrics to determine when the proxy has actually
accepted a connection from the test client, so that it can advance to
the next state; it was necessary to re-enable transport metrics for this
to work (see PR #530).

Additionally, there was a subtle change in the test-support client's
behavior that needed to be resolved as well: the test relied on
`Client::request_async` *eagerly* causing the test client to connect to
the proxy when called, rather than when the returned future is polled.
Switching `request_async` to an `async fn` broke this property — the
*entire* body of the function, even the synchronous parts, was run only
when the `async fn` future is polled. This means that the connection was
not opened when the test expected it to be, and the subsequent assertion
that it was open always failed.

Therefore, I changed `request_async` back into a synchronous function
returning `impl Future`. This is identical at the callsite (it still
returns something that you can `.await`), but calling the function now
once again eagerly tells the test client background thread to open a
connection. This fixes the test.

Depends on #530.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>